requirements-production.txt
# Core FastAPI and ASGI
fastapi==0.115.12
uvicorn==0.34.2
uvloop==0.21.0
python-multipart

# Socket.IO
python-socketio
aiofiles

# LangChain Core
langchain==0.3.24
langchain-anthropic==0.3.12
langchain-chroma==0.2.3
langchain-community==0.3.22
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-text-splitters==0.3.8

# LangGraph
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-checkpoint-postgres==2.0.21
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.33

# AI/ML
openai==1.75.0
anthropic==0.50.0
tiktoken==0.9.0

# Vector Database
chromadb==0.6.3
chroma-hnswlib==0.7.6

# Database
supabase==2.15.1
psycopg==3.2.6
SQLAlchemy==2.0.40

# HTTP and Networking
httpx==0.28.1
requests==2.32.3
aiohttp==3.11.18

# Data Processing
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
numpy==2.2.5

# Utilities
tenacity==9.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
tqdm==4.67.1 